Add localization types package
