+++ apps/ui/src/app/[lang]/second-page/Second.tsx
@@ -18,9 +18,9 @@
  */
 
 import Link from 'next/link';
+import { ValidLanguage } from 'common/@types/localization';
 
 import { getTranslation } from '@/app/i18n';
-import { ValidLanguage } from '@/app/i18n/settings';
 
 const Second = async ({ lang }: { lang: ValidLanguage }) => {
   const translate = await getTranslation(lang, 'second-page');
