+++ packages/common/@types/localization/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "localization",
+  "version": "1.0.0",
+  "description": "Type definitions for localization",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "zod": "^3.21.4"
+  }
+}
