+++ packages/common/@types/localization/src/index.ts
@@ -1,0 +1,5 @@
+import * as z from 'zod';
+
+const LANGUAGES = ['en', 'fr'] as const;
+export const ValidLanguage = z.enum(LANGUAGES);
+export type ValidLanguage = z.infer<typeof ValidLanguage>;
