+++ packages/common/@types/localization/tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "extends": "../../../../tsconfig.base.json",
+    "compilerOptions": {
+      "outDir": "dist",
+      "rootDir": "src",
+      "baseUrl": ".",
+      "declaration": true,
+    },
+    "include": ["src"]
+}
+
